分页等

+++ api/paginate.go
@@ -3,14 +3,14 @@ package api
 import "github.com/gin-gonic/gin"
 
 type PageMeta struct {
-	Total uint `json:"total"`
+	Total int64 `json:"total"`
 }
 type Pagination struct {
 	Data interface{} `json:"data"`
 	Meta PageMeta `json:"meta"`
 }
 
-func Page(ctx *gin.Context,data interface{},total uint)  {
+func Page(ctx *gin.Context,data interface{},total int64)  {
 	Success(ctx,Pagination{
 		Data: data,
 		Meta: PageMeta{
